Set an autocomplete for requiredFields and email
There was no autocomplete attribute set for the required fields, which made it hard for the browser / assistive technology to set the correct values for these fields. The default autocomplete values for the requiredFields are set in the config/user.js file, but are overwriteable through the client config in the database. See the updated README for an example.

The email field in the url/login view has also been given an autocomplete attribute with the value of "email" to give more clarity.